Extract click chart in to component
