+++ frontend/styles/Dashboard.module.css
@@ -9,36 +9,3 @@
   padding-left: 4rem;
   padding-right: 4rem;
 }
-
-.chart {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  margin-bottom: 4rem;
-  padding: 1rem 2rem;
-  border: #333 1px solid;
-  border-radius: 1rem;
-}
-
-.chartTitle {
-  font-size: 1.8rem;
-  margin-bottom: 2rem;
-}
-
-.googleChartContainer {
-  width: 100%;
-  height: 450px;
-}
-
-.googleChartContainer :global(div.google-visualization-tooltip) {
-  background-color: #222;
-  border: #555 2px solid;
-  border-radius: 0.5rem;
-  box-shadow: none;
-  -webkit-box-shadow: none;
-}
-
-.googleChartContainer :global(div.google-visualization-tooltip > ul > li > span) {
-  color: white !important;
-}
