+++ frontend/styles/components/DashboardClickChart.module.css
@@ -1,0 +1,37 @@
+.chart {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  margin-bottom: 4rem;
+  padding: 1rem 2rem;
+  border: #333 1px solid;
+  border-radius: 1rem;
+}
+
+.chart:last-child {
+  margin-bottom: 0;
+}
+
+.chartTitle {
+  font-size: 1.8rem;
+  margin-bottom: 2rem;
+  font-weight: bold;
+}
+
+.googleChartContainer {
+  width: 100%;
+  height: 450px;
+}
+
+.googleChartContainer :global(div.google-visualization-tooltip) {
+  background-color: #222;
+  border: #555 2px solid;
+  border-radius: 0.5rem;
+  box-shadow: none;
+  -webkit-box-shadow: none;
+}
+
+.googleChartContainer :global(div.google-visualization-tooltip > ul > li > span) {
+  color: white !important;
+}
